Uploaded level pattern generator and its respective input file.
* Walls are black.
* Tiles are white.
* Friendly intermediary coordinates are teal.
* Enemy intermediary coordinates are yellow.
* Overlapping friendly and enemy coordinates are grey.
   * Can't really do anything about this. Maybe they could be used as some sort of special item spawn or ignored?
* Friendly tower is dark blue.
* Enemy tower is orange.
* Friendly start is green.
   * Toggleable
* Enemy start is red.
   * Toggleable


In the json file, the grid size, overall number of rooms, path width, tower box width, spawn box width, and intermediary box width can be set. Whether the spawn boxes are generated or not can also be set (0 for No and 1 for Yes).


The json file and jupyter notebook have to be in the same location. The output image is also generated in the same location.