"Intermediary Towers" Spawning
Coordinates were not being cooperative, so I instead tried to do spawning based on colors for the main towers. While the main tower locations still aren't working, I accidentally found the colors required for intermediary spawning. Turns out rgb for unity is from 0 to 1 with floats instead of 0 to 255 with integers. The main towers are currently spawning in the intermediary locations. This will be changed later.